fix: Use float instead of Decimal for geocode response coordinates
Decimal values serialize to strings in JSON, causing the frontend form
to show "invalid numeric format" errors when coordinates are auto-filled
by the lookup function.

Changed GeocodeResponse schema and location_service to use float for
latitude/longitude, which serializes correctly as JSON numbers.

Issue #39 - Calendar Events feature (Phase 8)

+++ backend/src/schemas/location.py
@@ -109,8 +109,8 @@ class GeocodeResponse(BaseModel):
     state: Optional[str] = Field(default=None, description="State/province")
     country: Optional[str] = Field(default=None, description="Country name")
     postal_code: Optional[str] = Field(default=None, description="ZIP/postal code")
-    latitude: Decimal = Field(..., description="Geocoded latitude")
-    longitude: Decimal = Field(..., description="Geocoded longitude")
+    latitude: float = Field(..., description="Geocoded latitude")
+    longitude: float = Field(..., description="Geocoded longitude")
     timezone: Optional[str] = Field(default=None, description="IANA timezone identifier")
 
     model_config = {
